Add allowBlank to Dropdown
Since @nickw was having issues with the dropdown not defaulting to the
first entry if the value was null (even though the options had a null value in it)
this seems like a good time to add an option to default to the first element of the supplied options
(the prop is called allowBlank) in case the value is not found in the options.

The handling of null/undefined values has been fixed too and if there actually is a value of null
in the options it is correctly selected now.