feat(input): prevent insertion of characters if maxLength is reached
Previously we'd let the user insert a character, and then, if the resulting string was to long remove it again (through onChange). That creates unnessesary DOM modifications and store updates, and more importantly, diverges from the behavior of the native maxlength attribute of <textarea>.

Now we prevent the characters from beeing inserted in the first place if the length reaches maxlenght.